Fix error in getItemsInCommonWithActionTo to prevent jiggling perpetual loader
The problem is authenticator keeps running and causing ComparingContainer to
re-render. This happens because both Authenticator and ComparingContainer
look to the fetching status state to each load a loader.  When
The ComparingContainer fetches, updates the fetching state and
renders the loader, that causes Authenticator to load its Loader, which
then causes ComparingContainer to rerender and load it's Loader,
in a vicious cycle.  Hopefully the solution is to have only one
Loader component somewhere that renders children.  Or different
fetching states...